Avoid duplicating numpy arrays
Fixes #135

Instead of using a UUID for the name of the numpy files, use their hash.
That way, when the same array already exists, we can use it instead of
storing another copy.

Implementation

I started out with using hashlib.sha256 instead of delegating hashing to
joblib. However, this caused several problems:

- sha256 only worked for C-contiguous arrays
- the sha256 of np.array(0), np.array([0]), np.array([[0]]) etc. were
  all the same, which prevents us from restoring the arrays to their
  original shape

I did not try the other hashlib algorithms. Both drawbacks could be
worked around (store the contiguity and shape and restore them after
loading), but since joblib is already a dependency, I just went with it.

The disadvantage of joblib.hash is that it uses either md5 or sha1, both
of which are not cryptographically secure anymore. However, for this use
case, that should not be a necessity, since the hash only serves as a
cache key.

+++ skops/io/_numpy.py
@@ -1,7 +1,8 @@
 import io
+import os
 from pathlib import Path
-from uuid import uuid4
 
+import joblib
 import numpy as np
 
 from ._general import function_get_instance
@@ -16,11 +17,13 @@ def ndarray_get_state(obj, dst):
     }
 
     try:
-        f_name = f"{uuid4()}.npy"
-        with open(Path(dst) / f_name, "wb") as f:
-            np.save(f, obj, allow_pickle=False)
-            res["type"] = "numpy"
-            res["file"] = f_name
+        digest = joblib.hash(obj)
+        f_name = f"{digest}.npy"
+        if not os.path.exists(f_name):
+            with open(Path(dst) / f_name, "wb") as f:
+                np.save(f, obj, allow_pickle=False)
+        res["type"] = "numpy"
+        res["file"] = f_name
     except ValueError:
         # Object arrays cannot be saved with allow_pickle=False, therefore we
         # convert them to a list and recursively call get_state on it.
